+++ src/python/pants/core/goals/multi_tool_goal_helper_test.py
@@ -9,11 +9,7 @@ from pathlib import Path
 import pytest
 
 from pants.core.goals.check import CheckResult, CheckResults
-from pants.core.goals.style_request import (
-    StyleRequest,
-    determine_specified_tool_names,
-    write_reports,
-)
+from pants.core.goals.multi_tool_goal_helper import determine_specified_tool_names, write_reports
 from pants.core.util_rules.distdir import DistDir
 from pants.engine.fs import EMPTY_DIGEST, Workspace
 from pants.testutil.rule_runner import RuleRunner
@@ -21,7 +17,7 @@ from pants.util.strutil import softwrap
 
 
 def test_determine_specified_tool_names() -> None:
-    class StyleReq(StyleRequest):
+    class StyleReq:
         name = "my-tool"
 
     with pytest.raises(ValueError) as exc:
